hyprland: use rofi-rbw directly
Previously, it was just calling a script that then called the actual
program. The difference in execution time shouldn't be gigantic, but
it's still a bit of unnecessary indirection. Doing something like it
might be useful as a generic interface for commands, but since I'm the
only one using this there's little benefit to it, as I'm always aware
of what programs I'm using for each task.