uninstall code analysis tools
they're most likely defined in the flake anyway, so there'll be no need for a second declaration.

Additionally, the versions defined in this config file and the flake might differ, which
may lead to differences in formatting. And if not defined in the flake, then using them
without consulting the relevant team members on each project